Handle optional fields to dcb input file.
Apparently (see #5), not all of the fields on the TransCAD dcb are mandatory. This can cause a failure if a particular bin is missing some of the fields. 

We may need to test and see what happens if one of the fields is missing but another is not. But for now, it seems that none of the optional fields are passed on to `get_df_from_binary()`.