Update get_df_from_binary.cpp
R data types doesn't support SHORT and FLOAT (it refers to INTEGER and NUMERIC). So, instead of directly declaring IntergetVector or NumericVector to store SHORT or FLOAT values, assign them to interim primitive types and then pass them to respective vectors. This is not my favorite style and may revise once I have enumerated a new datatype to handle this. This should fix #15.